A collection of rat ligand-receptor interactions.

This dataset contains a data.table of curated rat ligand-receptor interactions as well as related annotations (GO Terms, KEGG Pathways) and metadata.

Usage

data(LRI_rat)

Format

A list with the following items:

LRI_curated: a data.table of curated LRIs

LRI_curated_GO: a data.table with GO terms attached to curated LRI

LRI_curated_KEGG: a data.table with KEGG pathways attached to curated LRIs

LRI_retrieved_dates: dates at which data have been retrieved from the seven external databases

LRI_retrieved_from: paths or packages from where data have been retrieved

LRI_biomart_ensembl_version: version of ensembl used for GO annotation and orthology conversion

Details

The dataset has been built internally in scDiffCom according to scDiffCom:::build_LRI(species = "rat"). The LRIs have been retrieved from seven databases (see References). Note that only curated LRIs have been kept.

References

CellChat (PMID: 33597522), CellPhoneDB (PMID: 32103204), CellTalkDB (PMID: 33147626), connectomeDB2020 (PMID: 33024107), ICELLNET (PMID: 33597528), NicheNet (PMID: 31819264), SingleCellSignalR (PMID: 32196115)