Display top over-represented keywords from a category of interest

Plot a graph that shows the top over-represented terms of a given category for a given regulation. Terms are ordered by their ORA scores, computed from their odds ratios and adjusted p-values.

PlotORA(
  object,
  category,
  regulation = c("UP", "DOWN", "FLAT"),
  max_terms_show = 20,
  GO_aspect = c("biological_process", "molecular_function", "cellular_component"),
  OR_threshold = 1,
  bh_p_value_threshold = 0.05
)

# S4 method for scDiffCom
PlotORA(
  object,
  category,
  regulation = c("UP", "DOWN", "FLAT"),
  max_terms_show = 20,
  GO_aspect = c("biological_process", "molecular_function", "cellular_component"),
  OR_threshold = 1,
  bh_p_value_threshold = 0.05
)

Arguments

object

scDiffCom object

category

ORA category to display. Must be the name of one of the category present in ora_table.

regulation

ORA regulation to display. Can be either UP (default), DOWN or FLAT.

max_terms_show

Maximum number of terms to display. Default is 20.

GO_aspect

Name of the GO aspect to display when category == "GO_TERMS". Can be either biological_process ( default), molecular_function or cellular_component.

OR_threshold

Only the terms with an odds ratio above this threshold will be displayed. Default is 1, meaning no filtering is performed.

bh_p_value_threshold

Only the terms with an adjusted p-value below this threshold (and always below 0.05) will be displayed. Default is 0.05.

Value

A ggplot object.

Details

The ORA score is computed as the product between log2(odds ratio) and -log10(adj. p-value).