Run over-representation analysis

Perform over-representation analysis (ORA) on a scDiffCom object, with the possibility to define new categories in addition to the standard ones supported by default.

Usage

RunORA(
  object,
  categories = c("LRI", "LIGAND_COMPLEX", "RECEPTOR_COMPLEX", "ER_CELLTYPES",
    "EMITTER_CELLTYPE", "RECEIVER_CELLTYPE", "GO_TERMS", "KEGG_PWS"),
  extra_annotations = NULL,
  overwrite = TRUE,
  verbose = TRUE
)

# S4 method for scDiffCom
RunORA(
  object,
  categories = c("LRI", "LIGAND_COMPLEX", "RECEPTOR_COMPLEX", "ER_CELLTYPES",
    "EMITTER_CELLTYPE", "RECEIVER_CELLTYPE", "GO_TERMS", "KEGG_PWS"),
  extra_annotations = NULL,
  overwrite = TRUE,
  verbose = TRUE
)

Arguments

object

scDiffCom object

categories

Names of the standard categories on which to perform ORA. Default is all standard categories, namely c("LRI", "LIGAND_COMPLEX", "RECEPTOR_COMPLEX", "ER_CELLTYPES", "EMITTER_CELLTYPE", "RECEIVER_CELLTYPE", "GO_TERMS", "KEGG_PWS")

extra_annotations

Convenience parameter to perform ORA on user-defined non-standard categories. If NULL (default), ORA is performed only on standard categories from categories. Otherwise it must be a list of data.tables or data.frames (see Details).

overwrite

If TRUE (default), previous results are overwritten in case they correspond to a category passed in categories.

verbose

If TRUE (default), progress messages are printed.

Value

A scDiffCom object with updated slot ora_table.

Details

Additional user-defined ORA categories can be added via the parameter extra_annotations. The data.frames or data.tables in this list must have exactly two columns that indicates a relationship between values from a standard category (first column) to values of the new category (second column). As a typical example, this vignette shows how to perform ORA on cell type families attached to each cell type.